extract_images_from_pdf/requirements.txt
Pillow==10.0.1
PyMuPDF==1.20.2
